UniQ2025/requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
appdirs==1.4.4
astunparse==1.6.3
attrs==25.3.0
Authlib==1.5.2
autograd==1.7.0
autoray==0.7.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cirq==1.5.0
cirq-aqt==1.5.0
cirq-core==1.5.0
cirq-google==1.5.0
cirq-ionq==1.5.0
cirq-pasqal==1.5.0
cirq-web==1.5.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
diastatic-malt==2.15.2
dill==0.3.9
dnspython==2.7.0
dparse==0.6.4
duet==0.2.9
email_validator==2.2.0
filelock==3.16.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Security==5.6.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
fonttools==4.55.8
gast==0.6.0
google-api-core==2.24.2
google-auth==2.38.0
google-pasta==0.2.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h5py==3.13.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
nltk==3.9.1
numpy==2.0.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pbr==6.1.0
PennyLane==0.40.0
PennyLane_Lightning==0.40.0
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.1
pylatexenc==2.10
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
qiskit==1.3.2
qiskit-aer==0.16.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
rustworkx==0.16.0
safety==3.3.1
safety-schemas==0.0.11
scipy==1.15.1
scipy-openblas32==0.3.29.0.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
stevedore==5.4.0
symengine==0.13.0
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
tomlkit==0.13.2
tqdm==4.67.1
typedunits==0.0.1.dev20250404020538
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
WTForms==3.2.1

streamlit 
psycopg2-binary 
sentry-sdk 
websockets 
fastapi
uvicorn
requests

